Backend/app/requirements.txt
fastapi
uvicorn
databases
sqlalchemy
pydantic
psycopg2-binary
python-dotenv